Modified CSS files
Minor edits: Changed the color buttons to a darker shade of red so its not the same color as the container on login and sign up buttons. Hovering over, transitions to black. Input fields is lighter. Class buttons and enter button now have shadows and a border. Instead of "hello,  username" now says "Welcome, username"